+++ src/TaylorSeries/TaylorSeries.jl
@@ -9,7 +9,7 @@ see also [`HomogeneousPolynomial`](@ref).
 """
 module Taylor
 
-
+using ..ComputationalGraphs
 #using Markdown
 
 
@@ -26,7 +26,7 @@ export get_order, get_numvars,
     get_variable_names, get_variable_symbols,
     # jacobian, hessian, jacobian!, hessian!,
     displayBigO, use_show_default,
-    getcoeff
+    getcoeff, getderivative, taylor_factorial
 # function __init__()
 #     @static if !isdefined(Base, :get_extension)
 #         @require IntervalArithmetic = "d1acc4aa-44c8-5952-acd4-ba5d80a2a253" begin
